pyproject.toml
[project]
name = "rl-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Valter Schütz", email = "valterschutz@proton.me" }
]
requires-python = ">=3.13"
dependencies = [
    "torchrl>=0.9.2",
]

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "pytest>=8.4.1",
]
